% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_from_post.R
\name{calc_diff_from_post}
\alias{calc_diff_from_post}
\title{calculate difference by simulating from the posterior}
\usage{
calc_diff_from_post(
  m,
  newdata,
  ci = 0.95,
  use_relative_diff = FALSE,
  nrep = 1000
)
}
\arguments{
\item{m}{model object}

\item{newdata}{data.frame nrow = 2 defining two points to compare}

\item{ci}{range of confidence interval}

\item{use_relative_diff}{provide estimates as a relative difference, otherwise
presented as an absolute difference}

\item{nrep}{number of samples of posterior to compute ci}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate the variance between a baseline prediction
and subsequent prediction in a GAM model using samples from the
posterior.
}
