% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_from_post.R
\name{calc_diff_vec_from_post}
\alias{calc_diff_vec_from_post}
\title{calculate difference of set of vectors by bootstrapping}
\usage{
calc_diff_vec_from_post(m, newdata, ci = 0.95, nrep = 1000)
}
\arguments{
\item{m}{model object}

\item{newdata}{data.frame nrow >= 2 defining points to compare to the first
row in the data frame}

\item{ci}{range of confidence interval}

\item{nrep}{number of samples of posterior to compute ci}
}
\description{
calculate difference of set of vectors by bootstrapping
}
