% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_from_bootstrap.R
\name{calc_generic_vector_from_bootstrap}
\alias{calc_generic_vector_from_bootstrap}
\title{Calculate difference of vector of model inputs by bootstrapping original
data}
\usage{
calc_generic_vector_from_bootstrap(
  m,
  newdata,
  U = NULL,
  ci = 0.95,
  use_relative_diff = FALSE,
  nrep = 100
)
}
\arguments{
\item{m}{model object}

\item{newdata}{data}

\item{U}{difference matrix. defines the vector of differences being computed}

\item{ci}{range of confidence interval}

\item{nrep}{number of samples of posterior to compute ci}
}
\value{
An \code{list} object. The output has the following elements:
\itemize{
\item m - mean
\item lc - lower confidence interval
\item uc - upper confidence interval
\item total_n - total number of samples in bootstrap
}
}
\description{
Calculate difference of vector of model inputs by bootstrapping original
data
}
