% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_from_post.R
\name{calc_generic_vector_from_post}
\alias{calc_generic_vector_from_post}
\title{calculate difference of set of vectors by bootstrapping}
\usage{
calc_generic_vector_from_post(
  m,
  newdata,
  U = NULL,
  ci = 0.95,
  use_relative_diff = FALSE,
  nrep = 1000
)
}
\arguments{
\item{m}{model object}

\item{newdata}{data.frame nrow >= 2 defining points to compare to the first
row in the data frame}

\item{U}{difference matrix. defines the vector of differences being computed}

\item{ci}{range of confidence interval}

\item{use_relative_diff}{provide estimates as a relative difference, otherwise
presented as an absolute difference}

\item{nrep}{number of samples of posterior to compute ci}
}
\description{
calculate difference of set of vectors by bootstrapping
}
