% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_counterfactual_gam.R
\name{calc_sum_counterfactual_gam}
\alias{calc_sum_counterfactual_gam}
\title{Estimation of sum of model output}
\usage{
calc_sum_counterfactual_gam(
  m,
  baseline_data,
  counter_data = NULL,
  ci = 0.95,
  delta = TRUE,
  use_relative_diff = FALSE,
  use_post = FALSE,
  nrep = 1000
)
}
\arguments{
\item{m}{model object}

\item{baseline_data}{data.frame of baseline predictors}

\item{counter_data}{data.frame of counterfactual predictors. \code{counter_data}
cannot be \code{NULL}}

\item{ci}{range of confidence interval}

\item{delta}{Boolean determines whether to apply the delta approximation
to a log link function}

\item{use_relative_diff}{provide estimates as a relative difference, otherwise
presented as an absolute difference}

\item{use_post}{Use posterior to sample estimaton of means and confidence intervals}

\item{nrep}{number of samples used for posterior sampling. Only used if \code{use_post}
is \code{TRUE}}
}
\value{
list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculate the variance between a baseline sum of predictions
against a counterfactual sum of predictions in a \link{gam}
based on modified code from gam predict function help
}
