% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{input_check}
\alias{input_check}
\title{check the input type for \link{calc_generic_from_post} and \link{calc_generic_from_bootsrap}
functions}
\usage{
input_check(npreds, U)
}
\description{
check the input type for \link{calc_generic_from_post} and \link{calc_generic_from_bootsrap}
functions
}
