% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff_vector_gam.R
\name{calc_diff_vector_gam}
\alias{calc_diff_vector_gam}
\title{Estimate of differences}
\usage{
calc_diff_vector_gam(
  m,
  newdata,
  ci = 0.95,
  use_relative_diff = FALSE,
  method = "delta",
  nrep = 1000
)
}
\arguments{
\item{m}{model object}

\item{newdata}{data.frame nrow >= 2 defining points to compare to the first
row in the data frame}

\item{ci}{range of confidence interval}

\item{use_relative_diff}{provide estimates as a relative difference, otherwise
presented as an absolute difference}

\item{method}{Selects which method is used to estimate confidence intervals
and expectation. Can be one of the following:
\itemize{
\item{"delta"}{Implements delta approximation.}
\item{"posterior"}{Implements posterior sampling. Use \code{nrep} to specify number of bootstrap samples}
\item{"bootstrap"}{Implements Efron Bootstrap. Use \code{nrep} to specify number of bootstrap samples}
}}

\item{nrep}{number of samples used for posterior sampling. Only used if \code{use_post}
is \code{TRUE}}

\item{use_post}{Use posterior to sample estimaton of means and confidence intervals}
}
\value{
list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate the variance between a baseline prediction
and subsequent predictions in a gam model
}
