% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff_vector_gam.R
\name{calc_diff_vector_gam}
\alias{calc_diff_vector_gam}
\title{Estimate of differences}
\usage{
calc_diff_vector_gam(m, newdata, ci = 0.95, delta = TRUE)
}
\arguments{
\item{m}{model object}

\item{newdata}{data.frame nrow >= 2 defining points to compare to the first
row in the data frame}

\item{ci}{range of confidence interval}

\item{delta}{boolean determines whether to apply the delta approximation
to a log link function}
}
\value{
list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Calculate the variance between a baseline prediction
and subsequent predictions in a gam model
based on modified code from gam predict function help
}
